multiple warning fixes with gcc-apple
git-svn-id: http://svn.ta3d.org/yuni/trunk@2440 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/io/file/file.cpp
@@ -230,7 +230,7 @@ namespace File
 
 		// retrieve the file size in bytes
 		f.seekFromEndOfFile(0);
-		uint64 filesize = f.tell();
+		uint64 filesize = (uint64) f.tell();
 
 		if (filesize == 0)
 		{
@@ -261,6 +261,7 @@ namespace File
 			# endif
 			return errNone;
 		}
+
 		if (filesize > hardlimit or filesize > (uint) -10)
 			return errMemoryLimit;
